refactor: consolidate agent implementations
