+++ minecraft_crafter/agent.py
@@ -10,16 +10,52 @@ import sys
 # Add parent directory to path for imports
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
-from src.agents.crafter_agent.agent import create_crafter_agent
+import structlog
+from google.adk.agents import LlmAgent
+
+from minecraft_coordinator.callbacks import get_configured_callbacks
 from src.bridge.bridge_manager import BridgeManager
 from src.config import get_config, setup_google_ai_credentials
 from src.logging_config import setup_logging
 from src.minecraft_bot_controller import BotController
 from src.minecraft_data_service import MinecraftDataService
+from src.tools.mineflayer_tools import create_mineflayer_tools
+
+from .prompt import CRAFTER_PROMPT
 
 # Setup logging
 setup_logging()
 logger = logging.getLogger(__name__)
+
+
+def create_crafter_agent(bot_controller=None, mc_data_service=None) -> LlmAgent:
+    """Create the crafter agent with structured output.
+
+    Args:
+        bot_controller: BotController instance
+        mc_data_service: MinecraftDataService instance
+
+    Returns:
+        Configured crafter agent that returns results via output_key
+    """
+    # Get configured callbacks
+    callbacks = get_configured_callbacks()
+
+    # Register callbacks individually as per ADK API
+    crafter = LlmAgent(
+        name="CrafterAgent",
+        model="gemini-2.0-flash",
+        instruction=CRAFTER_PROMPT,
+        output_key="crafting_result",  # Structured output to state
+        tools=create_mineflayer_tools(bot_controller, mc_data_service),
+        **callbacks,  # Unpack callback dict to pass as individual parameters
+    )
+
+    # Add logger for callbacks to use
+    crafter._logger = structlog.get_logger(f"agents.{crafter.name}")
+
+    return crafter
+
 
 try:
     # Initialize configuration
