--- src/agents/__init__.py
@@ -1,4 +1,0 @@
-"""Agent modules for the Minecraft ADK Guild system."""
-
-# The agents are now created via factory functions, not classes
-# See individual agent modules for create_*_agent() functions
