--- src/agents/coordinator_agent/__init__.py
@@ -1,5 +1,0 @@
-"""Coordinator Agent module exports."""
-
-from .agent import create_coordinator_agent
-
-__all__ = ["create_coordinator_agent"]
