--- src/agents/crafter_agent/__init__.py
@@ -1,5 +1,0 @@
-"""Crafter Agent module exports."""
-
-from .agent import create_crafter_agent
-
-__all__ = ["create_crafter_agent"]
