--- src/agents/gatherer_agent/__init__.py
@@ -1,5 +1,0 @@
-"""Gatherer Agent module exports."""
-
-from .agent import create_gatherer_agent
-
-__all__ = ["create_gatherer_agent"]
