+++ src/tools/__init__.py
@@ -1,4 +1,3 @@
-from .agent_tools import create_crafter_tools, create_gatherer_tools
 from .mineflayer_tools import create_mineflayer_tools
 
-__all__ = ["create_mineflayer_tools", "create_gatherer_tools", "create_crafter_tools"]
+__all__ = ["create_mineflayer_tools"]
