+++ src/tools/mineflayer_tools.py
@@ -6,7 +6,6 @@ from typing import Any, Dict, List, Optional
 
 from google.adk.tools import ToolContext
 
-from ..agents.state_schema import StateKeys
 from ..logging_config import get_logger
 from ..minecraft_bot_controller import BotController
 from ..minecraft_data_service import MinecraftDataService
@@ -72,7 +71,7 @@ async def move_to(
 
         # Store movement start in state
         if tool_context and hasattr(tool_context, "state"):
-            tool_context.state[StateKeys.MOVEMENT_IN_PROGRESS] = {
+            tool_context.state["minecraft.movement_in_progress"] = {
                 "target": {"x": x, "y": y, "z": z},
                 "start_position": current_pos,
                 "start_time": __import__("time").time(),
@@ -90,13 +89,13 @@ async def move_to(
 
             # Update position in state
             if tool_context and hasattr(tool_context, "state"):
-                tool_context.state[StateKeys.MINECRAFT_POSITION] = {
+                tool_context.state["minecraft.position"] = {
                     "x": actual_pos["x"],
                     "y": actual_pos["y"],
                     "z": actual_pos["z"],
                     "timestamp": __import__("time").time(),
                 }
-                tool_context.state[StateKeys.MOVEMENT_IN_PROGRESS] = None
+                tool_context.state["minecraft.movement_in_progress"] = None
 
             return {
                 "status": "success",
@@ -358,7 +357,7 @@ async def get_movement_status(tool_context: Optional[ToolContext] = None) -> Dic
 
         # Check session state for movement tracking
         if tool_context and hasattr(tool_context, "state"):
-            movement_progress = tool_context.state.get(StateKeys.MOVEMENT_IN_PROGRESS)
+            movement_progress = tool_context.state.get("minecraft.movement_in_progress")
             if movement_progress:
                 # Calculate if we're still moving based on session state
                 current_time = __import__("time").time()
@@ -706,7 +705,7 @@ async def get_inventory(tool_context: Optional[ToolContext] = None) -> Dict[str,
 
         # Save structured inventory data to session state if tool_context is provided
         if tool_context and hasattr(tool_context, "state"):
-            tool_context.state[StateKeys.MINECRAFT_INVENTORY] = {
+            tool_context.state["minecraft.inventory"] = {
                 "items": enriched_items,
                 "summary": inventory_summary,
                 "statistics": result["statistics"],
@@ -723,7 +722,7 @@ async def get_inventory(tool_context: Optional[ToolContext] = None) -> Dict[str,
 
         # Save error state if tool_context is provided
         if tool_context and hasattr(tool_context, "state"):
-            tool_context.state[StateKeys.MINECRAFT_INVENTORY] = {
+            tool_context.state["minecraft.inventory"] = {
                 "error": str(e),
                 "timestamp": __import__("time").time(),
             }
